Run `npm compile` on F5 debug
Define a task in `tasks.json` and make that the default build task.
This task is executing the `npm compile` script.  Modify `launch.json`
to run this task before launching the extension.

Could have make the task NOT rely on npm and execute `tsc` directly
but it seems better to have it run the npm script since that is what
is presented in the `npm scripts` section of the UI.