Use ts-mocha for unit testing
Scaffolding for unit testing the Dictionary class.  Had to comment
out parts of the class that call on the vscode api.  It does not
appear that the api is available outside of the extension host, which
it makes sense that a call to a method using the api would throw an
error but vscode is not letting me instantiate the class if when there
is a vscode import.

Going to need to further break apart the logic of the dictionary and
the mechanism that gets the data for the dictionary.  This seems like
a good case for dependency injection so we can provide mock data for
testing (use a dummy class that returns an object literal).