Continue unit testing, refactor removeLeadingZeros
Wrote a test to make sure we can parse a variable like 'VC33'.  This
resulted in finding a bug where we assumed the code would be one letter
followed by one to three numbers.  This assumption was baked into the
method being used to remove leading zeros from the given code.

Ended up refactoring the removeLeadingZeros so it can be used for both
input codes and for keys in the settings object.