Add html file to aid with selection of scopes
Working on implemting an adaptive token feature.  We'll need to know
which tokens render as which color in various themes.  I've collected
the data in `themeData.js` by switching to each theme I have installed
and running the `Developer: Generate Color Theme From Current Settings`
command.  The `name` property is the display name of the theme that
generated the data. The `editorBackground` property is copied from
`colors.editorBackground` and `tokenColors` is a direct copy/paste.

This data is read by `colors.html` to show each scope the theme targets
and the color applied to the scope.

These files are simply to aid with design of the extension and will not
be included in the .visx package.