Provide a command to force colorization
Until this point we've been listening for changes to the color theme
and handling that event.  The handler includes a pre-flight check to
see if the grammar is already setup for the themeName found in settings.
When the grammar is not tailored to the current theme, a message is
presented to the user giving the option to update the grammar.

What happens if the user declines the update or if (during testing for
example) the grammar is already tailored to the active theme and we
just want to apply the latest changes?  These scenarios are addressed
by providing a command that forces the grammar to be re-written.

Code is going to need cleaned up a bit, tried to keep it broken apart
a bit to facilitate testing of the methods that modify the grammar.