Refresh inline comments on save and document change
This is a bug fix and a refactoring.  The registration of event
handlers was moved into the constructor, it doesn't need to be
called besides at instantiation.

A bug was created in 7811dd8d where the handler for
`onDidChangeActiveTextEditor` was moved to `onWillSaveTextDocument`.
We actually need to handle both events, otherwise, when you switch
to a new document the dictionary from the previous document remain
in memory until the new document is saved.