refine(dictionary.ts): Use more techniques from Webview example
When the command is called for the first time, create the webview
and keep a reference to it in the dictionary object.  The original
webview is revealed on each subsequent call.  The reference is
dropped when the webview is disposed.

Also, refactor the logic that generates the HTML into it's own
method.  This will need to be used when the dictionary is updated.

https://code.visualstudio.com/api/extension-guides/webview